+++ bin/faucet/src/server/pow.rs
@@ -8,7 +8,7 @@ use std::{
 use miden_objects::account::AccountId;
 use tokio::time::{Duration, interval};
 
-use super::challenge::{CHALLENGE_LIFETIME_SECONDS, Challenge};
+use super::challenge::Challenge;
 use crate::server::{ApiKey, get_pow::PowRequest, get_tokens::MintRequestError};
 
 /// The interval at which the challenge cache is cleaned up.
@@ -21,6 +21,7 @@ const CLEANUP_INTERVAL_SECONDS: u64 = 2;
 pub(crate) struct PoW {
     secret: [u8; 32],
     challenge_cache: Arc<Mutex<ChallengeCache>>,
+    challenge_lifetime: Duration,
     requests_per_difficulty_level: NonZeroUsize,
     initial_target_shift: u8,
 }
@@ -29,6 +30,7 @@ impl PoW {
     /// Creates a new `PoW` instance.
     pub fn new(
         secret: [u8; 32],
+        challenge_lifetime: Duration,
         requests_per_difficulty_level: NonZeroUsize,
         initial_target_shift: u8,
     ) -> Self {
@@ -37,12 +39,13 @@ impl PoW {
         // Start the cleanup task
         let cleanup_state = challenge_cache.clone();
         tokio::spawn(async move {
-            ChallengeCache::run_cleanup(cleanup_state).await;
+            ChallengeCache::run_cleanup(cleanup_state, challenge_lifetime).await;
         });
 
         Self {
             secret,
             challenge_cache,
+            challenge_lifetime,
             requests_per_difficulty_level,
             initial_target_shift,
         }
@@ -50,7 +53,7 @@ impl PoW {
 
     /// Generates a new challenge.
     pub fn build_challenge(&self, request: PowRequest) -> Challenge {
-        let timestamp = SystemTime::now()
+        let current_time = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("current timestamp should be greater than unix epoch")
             .as_secs();
@@ -59,7 +62,7 @@ impl PoW {
         Challenge::new(
             self.initial_target_shift,
             difficulty,
-            timestamp,
+            current_time,
             request.account_id,
             request.api_key,
             self.secret,
@@ -87,8 +90,7 @@ impl PoW {
     /// * The challenge is expired.
     /// * The challenge is invalid.
     /// * The challenge was already used.
-    /// * The account has already submitted a challenge recently (within the last
-    ///   [`CHALLENGE_LIFETIME_SECONDS`] seconds).
+    /// * The account has already submitted a challenge recently and it's not expired yet.
     ///
     /// # Panics
     /// Panics if the challenge cache lock is poisoned.
@@ -103,7 +105,7 @@ impl PoW {
         let challenge = Challenge::decode(challenge, self.secret)?;
 
         // Check timestamp validity
-        if challenge.is_expired(current_time) {
+        if challenge.is_expired(current_time, self.challenge_lifetime) {
             return Err(MintRequestError::ExpiredServerTimestamp(
                 challenge.timestamp,
                 current_time,
@@ -199,15 +201,15 @@ impl ChallengeCache {
     /// Cleanup expired challenges and update the number of challenges submitted per API key and
     /// account id.
     ///
-    /// Challenges are expired if they are older than [`CHALLENGE_LIFETIME_SECONDS`] seconds.
-    ///
     /// # Arguments
     /// * `current_time` - The current timestamp in seconds since the UNIX epoch.
+    /// * `challenge_lifetime` - The duration during which a challenge is valid.
     ///
     /// # Panics
     /// Panics if any expired challenge has no corresponding entries on the account or API key maps.
-    fn cleanup_expired_challenges(&mut self, current_time: u64) {
-        let limit_timestamp = current_time - CHALLENGE_LIFETIME_SECONDS;
+    fn cleanup_expired_challenges(&mut self, current_time: u64, challenge_lifetime: Duration) {
+        // Challenges older than this are expired.
+        let limit_timestamp = current_time - challenge_lifetime.as_secs();
 
         let valid_challenges = self.challenges.split_off(&limit_timestamp);
         let expired_challenges = std::mem::replace(&mut self.challenges, valid_challenges);
@@ -246,7 +248,7 @@ impl ChallengeCache {
     /// The cleanup task is responsible for removing expired challenges from the cache.
     /// It runs every minute and removes challenges that are no longer valid because of their
     /// timestamp.
-    pub async fn run_cleanup(cache: Arc<Mutex<Self>>) {
+    pub async fn run_cleanup(cache: Arc<Mutex<Self>>, challenge_lifetime: Duration) {
         let mut interval = interval(Duration::from_secs(CLEANUP_INTERVAL_SECONDS));
 
         loop {
@@ -258,7 +260,7 @@ impl ChallengeCache {
             cache
                 .lock()
                 .expect("challenge cache lock should not be poisoned")
-                .cleanup_expired_challenges(current_time);
+                .cleanup_expired_challenges(current_time, challenge_lifetime);
         }
     }
 }
@@ -287,7 +289,7 @@ mod tests {
     #[tokio::test]
     async fn test_pow_validation() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let pow = PoW::new(secret, Duration::from_secs(30), NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
@@ -311,7 +313,7 @@ mod tests {
     #[tokio::test]
     async fn test_timestamp_validation() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let pow = PoW::new(secret, Duration::from_secs(30), NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
@@ -326,7 +328,7 @@ mod tests {
             &api_key,
             &challenge.encode(),
             nonce,
-            current_time + CHALLENGE_LIFETIME_SECONDS + 1,
+            current_time + pow.challenge_lifetime.as_secs() + 1,
         );
         assert!(result.is_err());
 
@@ -339,7 +341,7 @@ mod tests {
     #[tokio::test]
     async fn account_id_is_rate_limited() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let pow = PoW::new(secret, Duration::from_secs(30), NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
@@ -368,7 +370,7 @@ mod tests {
     #[tokio::test]
     async fn submit_challenge_and_check_difficulty() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let pow = PoW::new(secret, Duration::from_secs(30), NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
@@ -389,7 +391,7 @@ mod tests {
     #[tokio::test]
     async fn test_cleanup_expired_challenges() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let pow = PoW::new(secret, Duration::from_secs(30), NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
@@ -398,13 +400,13 @@ mod tests {
         // build challenge manually with past timestamp to ensure that expires in 1 second
         let challenge = Challenge::from_parts(
             BigUint::from_bytes_be(&[0xff; 32]),
-            current_time - CHALLENGE_LIFETIME_SECONDS,
+            current_time - pow.challenge_lifetime.as_secs(),
             account_id,
             api_key.clone(),
             Challenge::compute_signature(
                 secret,
                 &BigUint::from_bytes_be(&[0xff; 32]),
-                current_time - CHALLENGE_LIFETIME_SECONDS,
+                current_time - pow.challenge_lifetime.as_secs(),
                 account_id,
                 &api_key.inner(),
             ),
@@ -436,7 +438,8 @@ mod tests {
     #[tokio::test]
     async fn difficulty_is_increased_with_load() {
         let secret = create_test_secret();
-        let pow = PoW::new(secret, NonZeroUsize::new(1).unwrap(), 0);
+        let challenge_lifetime = Duration::from_secs(30);
+        let pow = PoW::new(secret, challenge_lifetime, NonZeroUsize::new(1).unwrap(), 0);
         let mut rng = ChaCha20Rng::from_seed(rand::random());
         let api_key = ApiKey::generate(&mut rng);